+++ Sources/SkipUI/SkipUI/Containers/List.swift
@@ -903,30 +903,6 @@ final class ListItemModifier: RenderModifier {
         return ListItemModifier(background: background, separator: separator)
     }
 }
-
-final class BadgeModifier: RenderModifier {
-    let badge: Text?
-    let prominence: BadgeProminence?
-
-    init(badge: Text? = nil, prominence: BadgeProminence? = nil) {
-        self.badge = badge
-        self.prominence = prominence
-        super.init()
-    }
-
-    static func combined(for renderable: Renderable) -> BadgeModifier {
-        var badge: Text? = nil
-        var prominence: BadgeProminence? = nil
-        renderable.forEachModifier {
-            if let badgeModifier = $0 as? BadgeModifier {
-                badge = badge ?? badgeModifier.badge
-                prominence = prominence ?? badgeModifier.prominence
-            }
-            return nil
-        }
-        return BadgeModifier(badge: badge, prominence: prominence)
-    }
-}
 #endif
 
 #if false
